LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/scripts/underover-parameters-and-embellished-operator-2-expected.txt


FAIL StretchStackBottomShiftDown assert_approx_equals: munder: under shift expected 30 +/- 1 but got 0
FAIL StretchStackGapBelowMin assert_approx_equals: munder: under gap expected 110 +/- 1 but got 0
FAIL StretchStackTopShiftUp assert_approx_equals: mover: over shift expected 50 +/- 1 but got 0
FAIL StretchStackGapAboveMin assert_approx_equals: mover: over shift expected 70 +/- 1 but got 0
→
→

→
→

→
→

→
→
